# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Syllabyte
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface StudentProfile {
  /** @nullable */
  student_name?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  date_of_birth?: string | null;
  /** @nullable */
  college_name?: string | null;
  /** @nullable */
  course_name?: string | null;
  /** @nullable */
  sem?: number | null;
  /** @nullable */
  photo_url?: string | null;
  class_roll_no: string;
  /** @nullable */
  linked_email?: string | null;
}

export interface PhotoUpdate {
  photo_url: string;
}

export interface LinkEmailBody {
  email: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetProfileByEmailParams = {
email: string;
};